Project fixes
