+++ Calculate.Tests/CalculatorTests.cs
@@ -1,0 +1,11 @@
+namespace Calculate.Tests;
+
+
+[TestClass]
+public class CalculatorTests
+{
+    [TestMethod]
+    public void TestMethod1()
+    {
+    }
+}
