--- Calculate/Calculate.Tests/CalculatorTests.cs
@@ -1,11 +1,0 @@
-namespace Calculate.Tests;
-
-
-[TestClass]
-public class CalculatorTests
-{
-    [TestMethod]
-    public void TestMethod1()
-    {
-    }
-}
