--- Calculate/Calculate.sln
@@ -1,36 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.8.34408.163
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Calculate", "Calculate\Calculate.csproj", "{E80D6425-3474-416F-8CE4-365629CB06A9}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Calculate.Tests", "Calculate.Tests\Calculate.Tests.csproj", "{B135559F-67C6-4763-B327-F9367B787246}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{8F6A83B3-0007-4EAB-9DD1-D0B8C2C611A5}"
-	ProjectSection(SolutionItems) = preProject
-		Directory.Build.props = Directory.Build.props
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E80D6425-3474-416F-8CE4-365629CB06A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E80D6425-3474-416F-8CE4-365629CB06A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E80D6425-3474-416F-8CE4-365629CB06A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E80D6425-3474-416F-8CE4-365629CB06A9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B135559F-67C6-4763-B327-F9367B787246}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B135559F-67C6-4763-B327-F9367B787246}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B135559F-67C6-4763-B327-F9367B787246}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B135559F-67C6-4763-B327-F9367B787246}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {03FE1156-B81C-4DE0-907F-9FDAFBC272F9}
-	EndGlobalSection
-EndGlobal
