--- Calculate/Calculate/Calculate.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
-
-
-	</PropertyGroup>
-	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-		<NoWarn>1701;1702;CA1707</NoWarn>
-		<WarningsNotAsErrors>CA1707</WarningsNotAsErrors>
-	</PropertyGroup>
-	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-		<NoWarn>1701;1702;CA1707</NoWarn>
-		<WarningsNotAsErrors>CA1707</WarningsNotAsErrors>
-	</PropertyGroup>
-
-</Project>
